project/requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.5
django-heroku==0.3.1
gunicorn==20.0.4
psycopg2==2.8.6
pytz==2020.5
sqlparse==0.4.1
whitenoise==5.2.0
Pillow==2.2.1
django-rest-auth==0.9.5
django-rest-authtoken==2.1.3
django-rest-knox==4.1.0
djangorestframework-simplejwt==4.4.0
djangorestframework-jwt == 1.11.0
PyJWT == 1.7.1
django-filter==2.3.0
pipenv==2020.8.13
django-cors-headers == 3.7.0
